[core] add run sanity checks (fixes #6)
MsfParser will now check for the right number of pb times
and gold times before either returning a run or writing one.
If there are too few, they will be padded with zeroes.
Having too many is ignored, because currently mist will
only work with as many times as there are split names.