reimplement window resize handling
i'm not really sure how it behaved before, but now it will
increase the number of splits on screen if there are more, starting
at the top. if splits have been scrolled off the top of the screen, they
will be re-shown before splits scrolled off the bottom.

When shrinking the window, splits at the bottom are always removed.