[core] add TimeType enum to fix skipping issues
Previously, when a split was skipped, a time of 0 was stored for that split.
However, this caused issues with comparisons if the skip was then undone.
Adding the TimeType enum allows the time that a skipped split took to be
stored, so it can be used for calculations but also differentiated from
times that were taken by a normal split.